Added v4 for parking lot

+++ Questions/ParkingLot/V4/PaymentStrategy/PaymentStrategy.java
@@ -1,5 +1,5 @@
 package Questions.ParkingLot.V4.PaymentStrategy;
 
 public interface PaymentStrategy {
-    boolean pay(int amount);
+    boolean pay(double amount);
 }
